feat: Add functions for retrieving file icons and calculating file sizes
Added `getIconForLineFileExtension()` function to retrieve the icon corresponding to a given file extension. The function searches the `extension-associations` section in the assets file and returns the corresponding icon data string.

Also added `getSizeForAbstractFile()` function to calculate the size of a file or sum of sizes of all files within a folder, recursively. The function accepts a `TFile` or `TFolder` object as a parameter and calculates the size using recursion. If the object is a `TFile`, it returns the size of the file extracted from `file.stat.size`. If the object is a `TFolder`, it iterates through each child element and recursively calls `getSizeForAbstractFile()` to calculate their respective sizes.

NOTE: Please